Extract Context.getActivity() extention function.
