+++ Demo/src/main/java/com/paypal/android/utils/ContextExt.kt
@@ -1,0 +1,12 @@
+package com.paypal.android.utils
+
+import android.content.Context
+import android.content.ContextWrapper
+import androidx.appcompat.app.AppCompatActivity
+
+// Ref: https://stackoverflow.com/a/68423182
+fun Context.getActivity(): AppCompatActivity? = when (this) {
+    is AppCompatActivity -> this
+    is ContextWrapper -> baseContext.getActivity()
+    else -> null
+}
